feat(analyzer): add opt-in check for unhandled thrown exceptions
This commit introduces a new, optional check to the analyzer for detecting potentially unhandled exceptions.

When enabled, the analyzer will report an `unhandled-thrown-type` error if a function or method can throw an exception that is neither
caught within a `try-catch` block nor documented in a `@throws` tag.

This feature is **disabled by default** and can be enabled via the `check_throws` configuration option or the `--check-throws=true` command-line
flag.

Signed-off-by: azjezz <azjezz@protonmail.com>