feat(analyzer): add provider for `Psl\Regex\capture_groups`
This commit introduces a special return type provider for the `Psl\Regex\capture_groups` function.

This allows the analyzer to infer a precise shape for the returned `Psl\Type\TypeInterface`, based on the
capture group names provided to the function.

When `capture_groups` is called with a literal array of group names, the handler constructs a keyed array type
(`TKeyedArray`) that includes these names as keys. This precise type is then used as the generic parameter for
the returned `TypeInterface`.

Previously, the analyzer would fall back to a generic type, losing valuable information about the expected keys in the captured data.

closes #364

Signed-off-by: azjezz <azjezz@protonmail.com>