feat(reporting): add configurable pager support
This commit introduces a flexible and configurable pager for displaying output, improving the user experience for long reports.

The pager can be enabled in several ways:
  - Using the `--pager` or `--pager=true` command-line flag.
  - Setting `use-pager = true` in your `mago.toml` configuration file.

If the pager is enabled in the configuration file, you can disable it for a single command run by using the `--pager=false` flag.

When the pager is enabled, Mago will intelligently decide when to use it. It will not be used under the following conditions:

- The reporting target is not `stdout`.
- The reporting format does not support paging (e.g., `github`, `json`, `checkstyle`).
- The `stdout` is not an interactive terminal (i.e., the output is being piped to another command).

You can specify which pager application to use in two ways (the environment variable takes precedence):

  - Set the pager option in your mago.toml:
    ```toml
    pager = "delta"
    ```
  - Set the `MAGO_PAGER` environment variable:
    ```toml
    MAGO_PAGER=delta mago lint
    ```

related to #343

Signed-off-by: azjezz <azjezz@protonmail.com>